orchestrator/requirements.txt
dagster==1.12.22
dagster-webserver==1.12.22
dagster-graphql==1.12.22
dagster-postgres==0.28.22
dagster-docker==0.28.22
dagster-aws==0.28.22
dagster-pipes
dagster-dbt==0.28.22
dagster-pandas==0.28.22
dagster-pandera==0.28.22
pandera==0.26.1
dask==2024.4.1
dask-expr==1.0.11
distributed
oracledb==2.0.1
numpy<2.0.0
pandas==2.2.3
pyarrow
requests
playwright
requests_mock==1.11.0
cpi==2.0.5
datasets
python-dateutil
matplotlib
sqlalchemy==2.0.25
datasets==2.15.0
plotly==5.22.0
prophet==1.1.5
pydantic==2.11.1
pytest-mock==3.12.0
psycopg2-binary==2.9.10
pytz==2023.4
openpyxl==3.1.2
xlrd==2.0.1
urllib3==2.2.1
certifi==2024.6.2
tenacity==8.3.0
